[Chore] 컴포넌트 분리 및 css 1차 디자인
