+++ capstone/src/styles/App.css
@@ -1,0 +1,65 @@
+html, body, #root {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+  font-family: 'Noto Sans KR', sans-serif;
+  overflow: hidden; 
+}
+
+.app-container {
+  display: flex;
+  height: 100vh;
+  width: 100vw;
+}
+
+
+.search-panel {
+  width: 360px;
+  background-color: #f1f3f5;
+  border-right: 1px solid #ccc;
+  box-shadow: 2px 0 6px rgba(0, 0, 0, 0.05);
+  padding: 16px;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+}
+
+.search-panel h2 {
+  font-size: 17px;
+  font-weight: 600;
+  color: #111;
+  margin: 0 0 10px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.search-panel input {
+  width: 93%;
+  padding: 10px;
+  font-size: 14px;
+  margin-bottom: 10px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
+
+.search-panel button {
+  padding: 10px;
+  background-color: #007aff;
+  color: white;
+  border: none;
+  font-weight: bold;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.search-panel button:hover {
+  background-color: #0056d2;
+}
+
+
+.map-container {
+  flex: 1;
+  height: 100%;
+  width: 100%;
+}
